refactor(transformations): split into separate files and simplify inheritance
- Split transformation-steps.ts and transformations.ts
- Remove TransformationStepBase: V1 defines all fields, V2 extends V1
- Remove TransformationBase: V1 defines container + steps, V2 extends V1
- TransformationV1 is now arktype validator (was plain TS type)
- Factory functions return alias types (Transformation, TransformationStep)
- Inline migrateStepV1ToV2 into .pipe() callback

Schema shapes and migration behavior unchanged.
