+++ apps/whispering/src/lib/services/db/models/transformations.ts
@@ -1,59 +1,14 @@
 import { type } from 'arktype';
 import { nanoid } from 'nanoid/non-secure';
-import { TRANSFORMATION_STEP_TYPES } from '$lib/constants/database';
 import {
-	ANTHROPIC_INFERENCE_MODELS,
-	GOOGLE_INFERENCE_MODELS,
-	GROQ_INFERENCE_MODELS,
-	INFERENCE_PROVIDER_IDS,
-	OPENAI_INFERENCE_MODELS,
-} from '$lib/constants/inference';
-
-/**
- * The current version of the TransformationStep schema.
- * Increment this when adding new fields or making breaking changes.
- */
-const CURRENT_TRANSFORMATION_STEP_VERSION = 2 as const;
+	TransformationStep,
+	TransformationStepV1,
+	TransformationStepV2,
+} from './transformation-steps.js';
 
 // ============================================================================
 // VERSION 1 (FROZEN)
 // ============================================================================
-
-/**
- * V1: Original schema without Custom provider fields.
- * Old data has no version field, so we default to 1.
- *
- * FROZEN: Do not modify. This represents the historical V1 schema.
- */
-const TransformationStepV1 = type({
-	id: 'string',
-	type: type.enumerated(...TRANSFORMATION_STEP_TYPES),
-	'prompt_transform.inference.provider': type.enumerated(
-		...INFERENCE_PROVIDER_IDS,
-	),
-	'prompt_transform.inference.provider.OpenAI.model': type.enumerated(
-		...OPENAI_INFERENCE_MODELS,
-	),
-	'prompt_transform.inference.provider.Groq.model': type.enumerated(
-		...GROQ_INFERENCE_MODELS,
-	),
-	'prompt_transform.inference.provider.Anthropic.model': type.enumerated(
-		...ANTHROPIC_INFERENCE_MODELS,
-	),
-	'prompt_transform.inference.provider.Google.model': type.enumerated(
-		...GOOGLE_INFERENCE_MODELS,
-	),
-	// OpenRouter model is a free string (user can enter any model)
-	'prompt_transform.inference.provider.OpenRouter.model': 'string',
-	'prompt_transform.systemPromptTemplate': 'string',
-	'prompt_transform.userPromptTemplate': 'string',
-	'find_replace.findText': 'string',
-	'find_replace.replaceText': 'string',
-	'find_replace.useRegex': 'boolean',
-	version: '1 = 1',
-});
-
-export type TransformationStepV1 = typeof TransformationStepV1.infer;
 
 /**
  * Transformation type containing V1 steps (before Custom provider fields).
@@ -62,84 +17,37 @@ export type TransformationStepV1 = typeof TransformationStepV1.infer;
  * Note: The Transformation fields themselves are unchanged; only the step
  * schema differs between "V1" and "V2".
  */
-export type TransformationV1 = {
-	id: string;
-	title: string;
-	description: string;
-	createdAt: string;
-	updatedAt: string;
-	steps: TransformationStepV1[];
-};
+const TransformationV1 = type({
+	id: 'string',
+	title: 'string',
+	description: 'string',
+	createdAt: 'string',
+	updatedAt: 'string',
+	steps: [TransformationStepV1, '[]'],
+});
+
+export type TransformationV1 = typeof TransformationV1.infer;
 
 // ============================================================================
 // VERSION 2 (CURRENT)
 // ============================================================================
 
 /**
- * V2: Added Custom provider fields for local LLM endpoints.
- * Extends V1 with:
- * - Custom.model: Model name for custom endpoints
- * - Custom.baseUrl: Per-step base URL (falls back to global setting)
- *
- * CURRENT VERSION: This is the latest schema.
- */
-const TransformationStepV2 = TransformationStepV1.merge({
-	version: '2',
-	/** Custom provider for local LLM endpoints (Ollama, LM Studio, llama.cpp, etc.) */
-	'prompt_transform.inference.provider.Custom.model': 'string',
-	/**
-	 * Per-step base URL for custom endpoints. Allows different steps to use
-	 * different local services (e.g., Ollama on :11434, LM Studio on :1234).
-	 * Falls back to global `completion.Custom.baseUrl` setting if empty.
-	 */
-	'prompt_transform.inference.provider.Custom.baseUrl': 'string',
-});
-
-export type TransformationStepV2 = typeof TransformationStepV2.infer;
-
-/**
  * Current Transformation schema with V2 steps.
+ * Extends V1 by using V2 steps.
  *
  * The Transformation container fields (id, title, description, createdAt, updatedAt)
  * have not changed since V1. Only TransformationStep has versioning.
  */
-const TransformationV2 = type({
-	id: 'string',
-	title: 'string',
-	description: 'string',
-	createdAt: 'string',
-	updatedAt: 'string',
+const TransformationV2 = TransformationV1.merge({
 	steps: [TransformationStepV2, '[]'],
 });
 
 export type TransformationV2 = typeof TransformationV2.infer;
 
 // ============================================================================
-// MIGRATING VALIDATORS
+// MIGRATING VALIDATOR
 // ============================================================================
-// These accept any version and migrate to the latest (V2).
-// Use these when reading data that might be from an older schema version.
-// ============================================================================
-
-/**
- * TransformationStep validator with automatic migration.
- * Accepts V1 or V2 and always outputs V2.
- */
-export const TransformationStep = TransformationStepV1.or(
-	TransformationStepV2,
-).pipe((step): TransformationStepV2 => {
-	if (step.version === 1) {
-		return {
-			...step,
-			version: 2,
-			'prompt_transform.inference.provider.Custom.model': '',
-			'prompt_transform.inference.provider.Custom.baseUrl': '',
-		};
-	}
-	return step;
-});
-
-export type TransformationStep = TransformationStepV2;
 
 /**
  * Transformation validator with automatic step migration.
@@ -153,10 +61,10 @@ export const Transformation = TransformationV2.merge({
 export type Transformation = TransformationV2;
 
 // ============================================================================
-// FACTORY FUNCTIONS
+// FACTORY FUNCTION
 // ============================================================================
 
-export function generateDefaultTransformation(): TransformationV2 {
+export function generateDefaultTransformation(): Transformation {
 	const now = new Date().toISOString();
 	return {
 		id: nanoid(),
@@ -165,31 +73,5 @@ export function generateDefaultTransformation(): TransformationV2 {
 		steps: [],
 		createdAt: now,
 		updatedAt: now,
-	};
-}
-
-export function generateDefaultTransformationStep(): TransformationStepV2 {
-	return {
-		version: CURRENT_TRANSFORMATION_STEP_VERSION,
-		id: nanoid(),
-		type: 'prompt_transform',
-		'prompt_transform.inference.provider': 'Google',
-		'prompt_transform.inference.provider.OpenAI.model': 'gpt-4o',
-		'prompt_transform.inference.provider.Groq.model': 'llama-3.3-70b-versatile',
-		'prompt_transform.inference.provider.Anthropic.model': 'claude-sonnet-4-0',
-		'prompt_transform.inference.provider.Google.model': 'gemini-2.5-flash',
-		'prompt_transform.inference.provider.OpenRouter.model':
-			'mistralai/mixtral-8x7b',
-		// Empty strings for Custom provider - user must configure when switching to Custom
-		// baseUrl falls back to global setting in transformer.ts
-		'prompt_transform.inference.provider.Custom.model': '',
-		'prompt_transform.inference.provider.Custom.baseUrl': '',
-
-		'prompt_transform.systemPromptTemplate': '',
-		'prompt_transform.userPromptTemplate': '',
-
-		'find_replace.findText': '',
-		'find_replace.replaceText': '',
-		'find_replace.useRegex': false,
 	};
 }
